requirements.txt
streamlit==1.13.0
pandas==1.3.3
plotly==5.3.1
scikit-learn>=0.18
pillow ==9.0.1
numpy==1.21.5
protobuf==3.20.1
